+++ gpsdclient/__init__.py
@@ -1,1 +1,3 @@
 from .client import GPSDClient
+
+__all__ = ["GPSDClient"]