+++ gpsdclient/client.py
@@ -1,11 +1,15 @@
 """
 A simple and lightweight GPSD client.
 """
+
+from __future__ import annotations
+
 import json
 import re
 import socket
+from collections.abc import Iterable
 from datetime import datetime, timezone
-from typing import Any, Dict, Iterable, Union
+from typing import Any, Union
 
 # old versions of gpsd with NTRIP sources emit invalid json which contains trailing
 # commas. As the json strings emitted by gpsd are well known to not contain structures
@@ -16,24 +20,25 @@ REGEX_TRAILING_COMMAS = re.compile(r"\s*,\s*}")
 FilterType = Union[str, Iterable[str]]
 
 
-def parse_datetime(x: Any) -> Union[Any, datetime]:
+def parse_datetime(x: Any) -> Any | datetime:
     """
     tries to convert the input into a `datetime` object if possible.
     """
     try:
         if isinstance(x, float):
-            return datetime.utcfromtimestamp(x).replace(tzinfo=timezone.utc)
-        elif isinstance(x, str):
+            return datetime.fromtimestamp(x, tz=timezone.utc).replace(
+                tzinfo=timezone.utc
+            )
+        if isinstance(x, str):
             # time zone information can be omitted because gps always sends UTC.
-            result = datetime.strptime(x, "%Y-%m-%dT%H:%M:%S.%fZ")
-            result = result.replace(tzinfo=timezone.utc)
-            return result
+            result = datetime.strptime(x, "%Y-%m-%dT%H:%M:%S.%f%z")
+            return result.replace(tzinfo=timezone.utc)
     except ValueError:
         pass
     return x
 
 
-def create_filter_regex(reports: Union[str, Iterable[str]] = set()) -> str:
+def create_filter_regex(reports: str | Iterable[str] = set()) -> str:
     """
     Dynamically assemble a regular expression to match the given report classes.
     This way we don't need to parse the json to filter by report.
@@ -41,8 +46,8 @@ def create_filter_regex(reports: Union[str, Iterable[str]] = set()) -> str:
     if isinstance(reports, str):
         reports = reports.split(",")
     if reports:
-        classes = set(x.strip().upper() for x in reports)
-        return r'"class":\s?"(%s)"' % "|".join(classes)
+        classes = {x.strip().upper() for x in reports}
+        return r'"class":\s?"({})"'.format("|".join(classes))
     return r".*"
 
 
@@ -50,12 +55,14 @@ class GPSDClient:
     def __init__(
         self,
         host: str = "127.0.0.1",
-        port: Union[str, int] = "2947",
-        timeout: Union[float, int, None] = None,
+        port: str | int = "2947",
+        timeout: float | int | None = None,
+        want_pps: bool = False,
     ):
         self.host = host
         self.port = port
         self.timeout = timeout
+        self.want_pps = want_pps
         self.sock = None  # type: Any
 
     def gpsd_lines(self):
@@ -64,32 +71,45 @@ class GPSDClient:
             address=(self.host, int(self.port)),
             timeout=self.timeout,
         )
-        self.sock.send(b'?WATCH={"enable":true,"json":true}\n')
+        options = {"enable": True, "json": True}
+        if self.want_pps:
+            options["pps"] = True
+        self.sock.send(b"?WATCH=")
+        self.sock.send(
+            json.dumps(options, indent=None, separators=(",", ":")).encode("utf-8")
+        )
+        self.sock.send(b"\n")
         yield from self.sock.makefile("r", encoding="utf-8")
 
-    def json_stream(self, filter: FilterType = set()) -> Iterable[str]:
-        filter_regex = re.compile(create_filter_regex(filter))
+    def json_stream(self, filter_spec: FilterType | None = None) -> Iterable[str]:
+        if filter_spec is None:
+            filter_spec = set()
+        filter_regex = re.compile(create_filter_regex(filter_spec))
 
         expect_version_header = True
         for line in self.gpsd_lines():
-            answ = line.strip()
-            if answ:
-                if expect_version_header and not answ.startswith('{"class":"VERSION"'):
-                    raise EnvironmentError(
+            answer = line.strip()
+            if answer:
+                if expect_version_header and not answer.startswith(
+                    '{"class":"VERSION"'
+                ):
+                    raise OSError(
                         "No valid gpsd version header received. Instead received:\n"
-                        "%s...\n"
-                        "Are you sure you are connecting to gpsd?" % answ[:100]
+                        f"{answer[:100]}...\n"
+                        "Are you sure you are connecting to gpsd?"
                     )
                 expect_version_header = False
 
-                if not filter or filter_regex.search(answ):
-                    cleaned_json = REGEX_TRAILING_COMMAS.sub("}", answ)
+                if not filter_spec or filter_regex.search(answer):
+                    cleaned_json = REGEX_TRAILING_COMMAS.sub("}", answer)
                     yield cleaned_json
 
     def dict_stream(
-        self, *, convert_datetime: bool = True, filter: FilterType = set()
-    ) -> Iterable[Dict[str, Any]]:
-        for line in self.json_stream(filter=filter):
+        self, *, convert_datetime: bool = True, filter_spec: FilterType | None = None
+    ) -> Iterable[dict[str, Any]]:
+        if filter_spec is None:
+            filter_spec = set()
+        for line in self.json_stream(filter_spec=filter_spec):
             result = json.loads(line)
             if convert_datetime and "time" in result:
                 result["time"] = parse_datetime(result["time"])
@@ -101,7 +121,7 @@ class GPSDClient:
         self.sock = None
 
     def __str__(self):
-        return "<GPSDClient(host=%s, port=%s)>" % (self.host, self.port)
+        return f"<GPSDClient(host={self.host}, port={self.port})>"
 
     def __enter__(self):
         return self